app/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.38.2
psycopg2-binary==2.9.3
python-jose==3.3.0
passlib==1.7.4